Fix links when Unreal is "Run as Administrator"
When the editor is running with elevated privileges (e.g. you're running
it through an IDE that's started with "Run as Administrator"), clicking
links would always open new Unreal instances instead of opening in the
editor that was already running.

This would happen because the Mailslot that was created would not be
writable by hermes_urls.exe, since it was *not* running elevated. The
default ACL of it would only allow Administrator members to write. This
is fixed by explicitly creating a SECURITY_DESCRIPTOR that allows the
current user to write to the Mailslot.